revised config handling and some behaviors
