+++ services/activate-session-request.go
@@ -31,7 +31,7 @@ type ActivateSessionRequest struct {
 }
 
 // NewActivateSessionRequest creates a new ActivateSessionRequest.
-func NewActivateSessionRequest(reqHeader *RequestHeader, sig *SignatureData, certs []*SignedSoftwareCertificate, locales []string, userToken datatypes.UserIdentityToken, tokenSig *SignatureData) *ActivateSessionRequest {
+func NewActivateSessionRequest(reqHeader *RequestHeader, sig *SignatureData, locales []string, userToken datatypes.UserIdentityToken, tokenSig *SignatureData) *ActivateSessionRequest {
 	return &ActivateSessionRequest{
 		TypeID: datatypes.NewExpandedNodeID(
 			false, false,
@@ -42,7 +42,7 @@ func NewActivateSessionRequest(reqHeader *RequestHeader, sig *SignatureData, cer
 		),
 		RequestHeader:              reqHeader,
 		ClientSignature:            sig,
-		ClientSoftwareCertificates: NewSignedSoftwareCertificateArray(certs),
+		ClientSoftwareCertificates: &SignedSoftwareCertificateArray{ArraySize: 0},
 		LocaleIDs:                  datatypes.NewStringArray(locales),
 		UserIdentityToken:          datatypes.NewExtensionObject(0x01, userToken),
 		UserTokenSignature:         tokenSig,
