+++ services/activate-session-request_test.go
@@ -28,7 +28,7 @@ var activateSessionRequestCases = []struct {
 				time.Date(2018, time.August, 10, 23, 0, 0, 0, time.UTC),
 				1, 0, 0, "", NewNullAdditionalHeader(), nil,
 			),
-			NewSignatureData("", nil), nil, nil,
+			NewSignatureData("", nil), nil,
 			datatypes.NewAnonymousIdentityToken("anonymous"),
 			NewSignatureData("", nil),
 		),
