+++ services/create-session-response_test.go
@@ -30,7 +30,7 @@ var createSessionResponseCases = []struct {
 				0x08, 0x22, 0x87, 0x62, 0xba, 0x81, 0xe1, 0x11,
 				0xa6, 0x43, 0xf8, 0x77, 0x7b, 0xc6, 0x2f, 0xc8,
 			}),
-			6000000, nil, nil, nil,
+			6000000, nil, nil,
 			NewSignatureData(
 				"http://www.w3.org/2000/09/xmldsig#rsa-sha1",
 				nil,
