+++ services/signature-data.go
@@ -32,7 +32,10 @@ func NewSignatureData(algorithm string, signature []byte) *SignatureData {
 // Specification: Part4, Table 15 and Table 17 (serverSignature and clientSignature).
 func NewSignatureDataFrom(cert, nonce []byte) *SignatureData {
 	// TODO: add calculation here.
-	return &SignatureData{}
+	if cert == nil && nonce == nil {
+		return NewSignatureData("", nil)
+	}
+	return NewSignatureData("", nil)
 }
 
 // DecodeSignatureData decodes given bytes into SignatureData.
