+++ uasc/session.go
@@ -13,125 +13,9 @@ import (
 	"sync"
 	"time"
 
-	"github.com/wmnsk/gopcua/status"
-
 	"github.com/wmnsk/gopcua/datatypes"
-	"github.com/wmnsk/gopcua/errors"
 	"github.com/wmnsk/gopcua/services"
 )
-
-// SessionConfig is a set of common configurations used in Session.
-type SessionConfig struct {
-	// AuthenticationToken is the secret Session identifier used to verify that the request is
-	// associated with the Session. The SessionAuthenticationToken type is defined in 7.31.
-	AuthenticationToken datatypes.NodeID
-	// ClientDescription is the information that describes the Client application.
-	// The type ApplicationDescription is defined in 7.1.
-	ClientDescription *services.ApplicationDescription
-	// ServerEndpoints is the list of Endpoints that the Server supports.
-	// The Server shall return a set of EndpointDescriptions available for the serverUri
-	// specified in the request. The EndpointDescription type is defined in 7.10. The Client
-	// shall verify this list with the list from a DiscoveryEndpoint if it used a
-	// DiscoveryEndpoint to fetch the EndpointDescriptions.
-	// It is recommended that Servers only include the server.applicationUri, endpointUrl,
-	// securityMode, securityPolicyUri, userIdentityTokens, transportProfileUri and
-	// securityLevel with all other parameters set to null. Only the recommended
-	// parameters shall be verified by the client.
-	ServerEndpoints []*services.EndpointDescription
-	// LocaleIDs is the list of locale ids in priority order for localized strings. The first
-	// LocaleId in the list has the highest priority. If the Server returns a localized string
-	// to the Client, the Server shall return the translation with the highest priority that
-	// it can. If it does not have a translation for any of the locales identified in this list,
-	// then it shall return the string value that it has and include the locale id with the
-	// string. See Part 3 for more detail on locale ids. If the Client fails to specify at least
-	// one locale id, the Server shall use any that it has.
-	// This parameter only needs to be specified during the first call to ActivateSession during
-	// a single application Session. If it is not specified the Server shall keep using the
-	// current localeIds for the Session.
-	LocaleIDs []string
-	// UserIdentityToken is the credentials of the user associated with the Client application.
-	// The Server uses these credentials to determine whether the Client should be allowed to
-	// activate a Session and what resources the Client has access to during this Session.
-	// The UserIdentityToken is an extensible parameter type defined in 7.36.
-	// The EndpointDescription specifies what UserIdentityTokens the Server shall accept.
-	// Null or empty user token shall always be interpreted as anonymous.
-	UserIdentityToken datatypes.UserIdentityToken
-	// If the Client specified a user identity token that supports digital signatures, then it
-	// shall create a signature and pass it as this parameter. Otherwise the parameter is null.
-	// The SignatureAlgorithm depends on the identity token type.
-	// The SignatureData type is defined in 7.32.
-	UserTokenSignature *services.SignatureData
-	// If Session works as a client, SessionTimeout is the requested maximum number of milliseconds
-	// that a Session should remain open without activity. If the Client fails to issue a Service
-	// request within this interval, then the Server shall automatically terminate the Client Session.
-	// If Session works as a server, SessionTimeout is an actual maximum number of milliseconds
-	// that a Session shall remain open without activity. The Server should attempt to honour the
-	// Client request for this parameter,but may negotiate this value up or down to meet its own constraints.
-	SessionTimeout uint64
-	// mySignature is is the client/serverSignature expected to receive from the other endpoint.
-	// This parameter is automatically calculated and kept temporarily until being used to verify
-	// received client/serverSignature.
-	mySignature *services.SignatureData
-	// signatureToSend is the client/serverSignature defined in Part4, Table 15 and Table 17.
-	// This parameter is automatically calculated and kept temporarily until it is sent in next message.
-	signatureToSend *services.SignatureData
-}
-
-// NewClientSessionConfig creates a SessionConfig for client.
-func NewClientSessionConfig(locales []string, userToken datatypes.UserIdentityToken) *SessionConfig {
-	return &SessionConfig{
-		SessionTimeout: 0xffff,
-		ClientDescription: services.NewApplicationDescription(
-			"urn:gopcua:client", "urn:gopcua", "gopcua - OPC UA implementation in pure Golang",
-			services.AppTypeClient, "", "", []string{""},
-		),
-		LocaleIDs:          locales,
-		UserIdentityToken:  userToken,
-		UserTokenSignature: services.NewSignatureData("", nil),
-	}
-}
-
-// NewServerSessionConfig creates a new SessionConfigServer for server.
-func NewServerSessionConfig(secChan *SecureChannel) *SessionConfig {
-	return &SessionConfig{
-		AuthenticationToken: datatypes.NewFourByteNodeID(0, uint16(time.Now().UnixNano())),
-		SessionTimeout:      0xffff,
-		ServerEndpoints: []*services.EndpointDescription{
-			services.NewEndpointDescription(
-				secChan.LocalEndpoint(), services.NewApplicationDescription(
-					"urn:gopcua:client", "urn:gopcua", "gopcua - OPC UA implementation in pure Golang",
-					services.AppTypeServer, "", "", []string{""},
-				),
-				secChan.cfg.Certificate, secChan.cfg.SecurityMode, secChan.cfg.SecurityPolicyURI,
-				services.NewUserTokenPolicyArray(
-					[]*services.UserTokenPolicy{
-						services.NewUserTokenPolicy("", 0, "", "", ""),
-					},
-				), "", 0,
-			),
-		},
-	}
-}
-
-// validate validates SessionConfig. This is just to avoid crash. Strange values would be accepted for flexibility.
-func (s *SessionConfig) validate(appType string) error {
-	switch appType {
-	case "client":
-		return s.validateClientSessionConfig()
-	case "server":
-		return s.validateClientSessionConfig()
-	default:
-		return errors.New("invalid type. should be client or server")
-	}
-}
-
-func (s *SessionConfig) validateClientSessionConfig() error {
-	return nil
-}
-
-func (s *SessionConfig) validateServerSessionConfig() error {
-	return nil
-}
 
 // Session is an implementation of the net.Conn interface for Session in OPC UA Secure Conversation.
 //
@@ -399,9 +283,12 @@ func (s *Session) handleCreateSessionResponse(cs *services.CreateSessionResponse
 	switch s.state {
 	case cliStateCreateSessionSent:
 		if cs.ResponseHeader.ServiceResult == 0 {
+			/* XXX - should be handled properly when sign and encryption enabled.
 			if err := validateSignature(cs.ServerSignature, s.cfg.mySignature); err != nil {
 				s.errChan <- err
 			}
+			*/
+
 			s.secChan.reqHeader.AuthenticationToken = cs.AuthenticationToken
 			s.cfg.ServerEndpoints = cs.ServerEndpoints.EndpointDescriptions
 			s.cfg.SessionTimeout = cs.RevisedSessionTimeout
@@ -424,6 +311,7 @@ func (s *Session) handleActivateSessionRequest(as *services.ActivateSessionReque
 
 	switch s.state {
 	case srvStateSessionCreated, srvStateSessionActivated:
+		/* XXX - should be handled properly when sign and encryption enabled.
 		if err := validateSignature(as.ClientSignature, s.cfg.mySignature); err != nil {
 			if err := s.ActivateSessionResponse(status.BadUserSignatureInvalid); err != nil {
 				s.errChan <- err
@@ -432,6 +320,7 @@ func (s *Session) handleActivateSessionRequest(as *services.ActivateSessionReque
 			s.errChan <- err
 			return
 		}
+		*/
 
 		for _, str := range as.LocaleIDs.Strings {
 			s.cfg.LocaleIDs = append(s.cfg.LocaleIDs, str.Get())
@@ -506,22 +395,24 @@ func (s *Session) handleCloseSessionResponse(cs *services.CloseSessionResponse) 
 
 // CreateSessionRequest sends a CreateSessionRequest.
 func (s *Session) CreateSessionRequest() error {
-	s.secChan.reqHeader.RequestHandle++
-	s.secChan.reqHeader.Timestamp = time.Now()
-
 	nonce := make([]byte, 32)
 	if _, err := rand.Read(nonce); err != nil {
 		return err
 	}
+
+	s.secChan.reqHeader.RequestHandle++
+	s.secChan.reqHeader.Timestamp = time.Now()
 	csr, err := services.NewCreateSessionRequest(
 		s.secChan.reqHeader, s.cfg.ClientDescription, "", s.secChan.RemoteEndpoint(),
 		"gopcua-"+time.Now().String(), nonce, s.secChan.cfg.Certificate, 0, 0,
 	).Serialize()
 	if err != nil {
+		s.secChan.reqHeader.RequestHandle--
 		return err
 	}
 
 	if _, err := s.secChan.WriteService(csr); err != nil {
+		s.secChan.reqHeader.RequestHandle--
 		return err
 	}
 
@@ -532,8 +423,6 @@ func (s *Session) CreateSessionRequest() error {
 
 // CreateSessionResponse sends a CreateSessionResponse.
 func (s *Session) CreateSessionResponse() error {
-	s.secChan.resHeader.Timestamp = time.Now()
-
 	sid := make([]byte, 4)
 	if _, err := rand.Read(sid); err != nil {
 		return err
@@ -544,10 +433,11 @@ func (s *Session) CreateSessionResponse() error {
 		return err
 	}
 
+	s.secChan.resHeader.Timestamp = time.Now()
 	csr, err := services.NewCreateSessionResponse(
 		// XXX - Give AuthenticationToken as NodeID
 		s.secChan.resHeader, datatypes.NewNumericNodeID(0, sessID), s.cfg.AuthenticationToken, 0xffff,
-		nonce, s.secChan.cfg.Certificate, nil, s.cfg.signatureToSend, 0xffff, s.cfg.ServerEndpoints...,
+		nonce, s.secChan.cfg.Certificate, s.cfg.signatureToSend, 0xffff, s.cfg.ServerEndpoints...,
 	).Serialize()
 	if err != nil {
 		return err
@@ -566,16 +456,17 @@ func (s *Session) CreateSessionResponse() error {
 func (s *Session) ActivateSessionRequest() error {
 	s.secChan.reqHeader.RequestHandle++
 	s.secChan.reqHeader.Timestamp = time.Now()
-
 	asr, err := services.NewActivateSessionRequest(
-		s.secChan.reqHeader, s.cfg.signatureToSend, nil, s.cfg.LocaleIDs,
-		s.cfg.UserIdentityToken, s.cfg.UserTokenSignature,
+		s.secChan.reqHeader, s.cfg.signatureToSend, s.cfg.LocaleIDs, s.cfg.UserIdentityToken,
+		s.cfg.UserTokenSignature,
 	).Serialize()
 	if err != nil {
+		s.secChan.reqHeader.RequestHandle--
 		return err
 	}
 
 	if _, err := s.secChan.WriteService(asr); err != nil {
+		s.secChan.reqHeader.RequestHandle--
 		return err
 	}
 	return nil
@@ -583,12 +474,12 @@ func (s *Session) ActivateSessionRequest() error {
 
 // ActivateSessionResponse sends a ActivateSessionResponse.
 func (s *Session) ActivateSessionResponse(results ...uint32) error {
-	s.secChan.resHeader.Timestamp = time.Now()
-
 	nonce := make([]byte, 32)
 	if _, err := rand.Read(nonce); err != nil {
 		return err
 	}
+
+	s.secChan.resHeader.Timestamp = time.Now()
 	asr, err := services.NewActivateSessionResponse(
 		s.secChan.resHeader, nonce, results, []*services.DiagnosticInfo{
 			services.NewNullDiagnosticInfo(),
@@ -608,15 +499,16 @@ func (s *Session) ActivateSessionResponse(results ...uint32) error {
 func (s *Session) CloseSessionRequest(delete bool) error {
 	s.secChan.reqHeader.RequestHandle++
 	s.secChan.reqHeader.Timestamp = time.Now()
-
 	csr, err := services.NewCloseSessionRequest(
 		s.secChan.reqHeader, delete,
 	).Serialize()
 	if err != nil {
+		s.secChan.reqHeader.RequestHandle--
 		return err
 	}
 
 	if _, err := s.secChan.WriteService(csr); err != nil {
+		s.secChan.reqHeader.RequestHandle--
 		return err
 	}
 	return nil
@@ -625,7 +517,6 @@ func (s *Session) CloseSessionRequest(delete bool) error {
 // CloseSessionResponse sends a CloseSessionResponse.
 func (s *Session) CloseSessionResponse() error {
 	s.secChan.resHeader.Timestamp = time.Now()
-
 	csr, err := services.NewCloseSessionResponse(s.secChan.resHeader).Serialize()
 	if err != nil {
 		return err
@@ -641,15 +532,16 @@ func (s *Session) CloseSessionResponse() error {
 func (s *Session) ReadRequest(maxAge uint64, tsRet services.TimestampsToReturn, nodes ...*datatypes.ReadValueID) error {
 	s.secChan.reqHeader.RequestHandle++
 	s.secChan.reqHeader.Timestamp = time.Now()
-
 	rdr, err := services.NewReadRequest(
 		s.secChan.reqHeader, maxAge, tsRet, nodes...,
 	).Serialize()
 	if err != nil {
+		s.secChan.reqHeader.RequestHandle--
 		return err
 	}
 
 	if _, err := s.secChan.WriteService(rdr); err != nil {
+		s.secChan.reqHeader.RequestHandle--
 		return err
 	}
 	return nil
@@ -658,7 +550,6 @@ func (s *Session) ReadRequest(maxAge uint64, tsRet services.TimestampsToReturn, 
 // ReadResponse sends a ReadResponse.
 func (s *Session) ReadResponse(results ...*datatypes.DataValue) error {
 	s.secChan.resHeader.Timestamp = time.Now()
-
 	rdr, err := services.NewReadResponse(
 		s.secChan.resHeader, nil, results...,
 	).Serialize()
@@ -672,60 +563,11 @@ func (s *Session) ReadResponse(results ...*datatypes.DataValue) error {
 	return nil
 }
 
-type sessionState uint8
-
-const (
-	cliStateSessionClosed sessionState = iota
-	cliStateCreateSessionSent
-	cliStateSessionCreated
-	cliStateActivateSessionSent
-	cliStateSessionActivated
-	cliStateCloseSessionSent
-	srvStateSessionClosed
-	srvStateSessionCreated
-	srvStateActivateSessionSent
-	srvStateSessionActivated
-	srvStateCloseSessionSent
-)
-
-func (s sessionState) String() string {
-	switch s {
-	case cliStateSessionClosed:
-		return "cliStateSessionClosed"
-	case cliStateCreateSessionSent:
-		return "cliStateCreateSessionSent"
-	case cliStateSessionCreated:
-		return "cliStateSessionCreated"
-	case cliStateActivateSessionSent:
-		return "cliStateActivateSessionSent"
-	case cliStateSessionActivated:
-		return "cliStateSessionActivated"
-	case cliStateCloseSessionSent:
-		return "cliStateCloseSessionSent"
-	case srvStateSessionClosed:
-		return "srvStateSessionClosed"
-	case srvStateSessionCreated:
-		return "srvStateSessionCreated"
-	case srvStateActivateSessionSent:
-		return "srvStateActivateSessionSent"
-	case srvStateSessionActivated:
-		return "srvStateSessionActivated"
-	case srvStateCloseSessionSent:
-		return "srvStateCloseSessionSent"
-	default:
-		return ""
-	}
-}
-
-// Errors
-var (
-	ErrInvalidAuthenticationToken = errors.New("invalid AuthenticationToken")
-	ErrSessionNotActivated        = errors.New("session is not activated")
-	ErrInvalidSignatureAlgorithm  = errors.New("algorithm in signature doesn't match")
-	ErrInvalidSignatureData       = errors.New("signature is invalid")
-)
-
 func validateSignature(received, expected *services.SignatureData) error {
+	if received.Algorithm == nil || expected.Algorithm == nil {
+		return nil
+	}
+
 	if received.Algorithm.Get() != expected.Algorithm.Get() {
 		return ErrInvalidSignatureAlgorithm
 	}
